+++ typescript/infra/config/environments/mainnet3/warp/configGetters/getEthereumFlowCbBTCWarpConfig.ts
@@ -2,8 +2,8 @@ import { ethers } from 'ethers';
 
 import {
   ChainMap,
+  HypTokenRouterConfig,
   OwnableConfig,
-  TokenRouterConfig,
   TokenType,
 } from '@hyperlane-xyz/sdk';
 
@@ -12,27 +12,30 @@ import {
   RouterConfigWithoutOwner,
   tokens,
 } from '../../../../../src/config/warp.js';
-import { DEPLOYER } from '../../owners.js';
 
-// Keep on our deployer for now until we get an address from Flow
-const owner = DEPLOYER;
-const ownerConfig = getOwnerConfigForAddress(owner);
+// Flow team Safe
+const ethereumOwner = '0x58C3FB862a4F5f038C24F8506BE378e9415c5B6C';
+const ethereumOwnerConfig = getOwnerConfigForAddress(ethereumOwner);
+
+// Flow team Safe
+const flowOwner = '0xa507DFccA02727B46cBdC600C57E89b2b55E5330';
+const flowOwnerConfig = getOwnerConfigForAddress(flowOwner);
 
 export const getEthereumFlowCbBTCWarpConfig = async (
   routerConfig: ChainMap<RouterConfigWithoutOwner>,
   _abacusWorksEnvOwnerConfig: ChainMap<OwnableConfig>,
-): Promise<ChainMap<TokenRouterConfig>> => {
-  const ethereum: TokenRouterConfig = {
+): Promise<ChainMap<HypTokenRouterConfig>> => {
+  const ethereum: HypTokenRouterConfig = {
     ...routerConfig.ethereum,
-    ...ownerConfig,
+    ...ethereumOwnerConfig,
     type: TokenType.collateral,
     token: tokens.ethereum.cbBTC,
     interchainSecurityModule: ethers.constants.AddressZero,
   };
 
-  const flowmainnet: TokenRouterConfig = {
+  const flowmainnet: HypTokenRouterConfig = {
     ...routerConfig.flowmainnet,
-    ...ownerConfig,
+    ...flowOwnerConfig,
     type: TokenType.synthetic,
     interchainSecurityModule: ethers.constants.AddressZero,
   };