+++ typescript/sdk/src/ism/types.ts
@@ -14,19 +14,13 @@ import {
 } from '@hyperlane-xyz/core';
 import type { Address, Domain, ValueOf } from '@hyperlane-xyz/utils';
 
-import { OwnableConfig } from '../deploy/types.js';
-import { ChainMap } from '../types.js';
-
+import { ZHash } from '../metadata/customZodTypes.js';
 import {
-  ArbL2ToL1IsmConfigSchema,
-  IsmConfigSchema,
-  MultisigIsmConfigSchema,
-  OpStackIsmConfigSchema,
-  PausableIsmConfigSchema,
-  TestIsmConfigSchema,
-  TrustedRelayerIsmConfigSchema,
-  WeightedMultisigIsmConfigSchema,
-} from './schemas.js';
+  ChainMap,
+  OwnableConfig,
+  OwnableSchema,
+  PausableSchema,
+} from '../types.js';
 
 // this enum should match the IInterchainSecurityModule.sol enum
 // meant for the relayer
@@ -51,6 +45,7 @@ export enum IsmType {
   OP_STACK = 'opStackIsm',
   ROUTING = 'domainRoutingIsm',
   FALLBACK_ROUTING = 'defaultFallbackRoutingIsm',
+  ICA_ROUTING = 'icaRoutingIsm',
   AGGREGATION = 'staticAggregationIsm',
   STORAGE_AGGREGATION = 'storageAggregationIsm',
   MERKLE_ROOT_MULTISIG = 'merkleRootMultisigIsm',
@@ -76,8 +71,8 @@ export const MUTABLE_ISM_TYPE = [
 export function ismTypeToModuleType(ismType: IsmType): ModuleType {
   switch (ismType) {
     case IsmType.ROUTING:
-      return ModuleType.ROUTING;
     case IsmType.FALLBACK_ROUTING:
+    case IsmType.ICA_ROUTING:
       return ModuleType.ROUTING;
     case IsmType.AGGREGATION:
     case IsmType.STORAGE_AGGREGATION:
@@ -103,8 +98,13 @@ export function ismTypeToModuleType(ismType: IsmType): ModuleType {
   }
 }
 
+export type ValidatorConfig = {
+  address: Address;
+  alias: string;
+};
+
 export type MultisigConfig = {
-  validators: Array<Address>;
+  validators: Array<ValidatorConfig>;
   threshold: number;
 };
 
@@ -126,10 +126,20 @@ export type NullIsmConfig =
   | OpStackIsmConfig
   | TrustedRelayerIsmConfig;
 
-export type RoutingIsmConfig = OwnableConfig & {
-  type: IsmType.ROUTING | IsmType.FALLBACK_ROUTING;
-  domains: ChainMap<IsmConfig>;
+type BaseRoutingIsmConfig<
+  T extends IsmType.ROUTING | IsmType.FALLBACK_ROUTING | IsmType.ICA_ROUTING,
+> = {
+  type: T;
 };
+
+export type DomainRoutingIsmConfig = BaseRoutingIsmConfig<
+  IsmType.ROUTING | IsmType.FALLBACK_ROUTING
+> &
+  OwnableConfig & { domains: ChainMap<IsmConfig> };
+
+export type IcaRoutingIsmConfig = BaseRoutingIsmConfig<IsmType.ICA_ROUTING>;
+
+export type RoutingIsmConfig = IcaRoutingIsmConfig | DomainRoutingIsmConfig;
 
 export type AggregationIsmConfig = {
   type: IsmType.AGGREGATION | IsmType.STORAGE_AGGREGATION;
@@ -143,6 +153,7 @@ export type DeployedIsmType = {
   [IsmType.CUSTOM]: IInterchainSecurityModule;
   [IsmType.ROUTING]: IRoutingIsm;
   [IsmType.FALLBACK_ROUTING]: IRoutingIsm;
+  [IsmType.ICA_ROUTING]: IRoutingIsm;
   [IsmType.AGGREGATION]: IAggregationIsm;
   [IsmType.STORAGE_AGGREGATION]: IAggregationIsm;
   [IsmType.MERKLE_ROOT_MULTISIG]: IMultisigIsm;
@@ -167,3 +178,106 @@ export type RoutingIsmDelta = {
   owner?: Address; // is the owner different
   mailbox?: Address; // is the mailbox different (only for fallback routing)
 };
+
+const ValidatorInfoSchema = z.object({
+  signingAddress: ZHash,
+  weight: z.number(),
+});
+
+export const TestIsmConfigSchema = z.object({
+  type: z.literal(IsmType.TEST_ISM),
+});
+
+export const MultisigConfigSchema = z.object({
+  validators: z.array(ZHash),
+  threshold: z.number(),
+});
+
+export const WeightedMultisigConfigSchema = z.object({
+  validators: z.array(ValidatorInfoSchema),
+  thresholdWeight: z.number(),
+});
+
+export const TrustedRelayerIsmConfigSchema = z.object({
+  type: z.literal(IsmType.TRUSTED_RELAYER),
+  relayer: z.string(),
+});
+
+export const OpStackIsmConfigSchema = z.object({
+  type: z.literal(IsmType.OP_STACK),
+  origin: z.string(),
+  nativeBridge: z.string(),
+});
+
+export const ArbL2ToL1IsmConfigSchema = z.object({
+  type: z.literal(IsmType.ARB_L2_TO_L1),
+  bridge: z.string(),
+});
+
+export const PausableIsmConfigSchema = PausableSchema.and(
+  z.object({
+    type: z.literal(IsmType.PAUSABLE),
+  }),
+);
+
+export const MultisigIsmConfigSchema = MultisigConfigSchema.and(
+  z.object({
+    type: z.union([
+      z.literal(IsmType.MERKLE_ROOT_MULTISIG),
+      z.literal(IsmType.MESSAGE_ID_MULTISIG),
+      z.literal(IsmType.STORAGE_MERKLE_ROOT_MULTISIG),
+      z.literal(IsmType.STORAGE_MESSAGE_ID_MULTISIG),
+    ]),
+  }),
+);
+
+export const WeightedMultisigIsmConfigSchema = WeightedMultisigConfigSchema.and(
+  z.object({
+    type: z.union([
+      z.literal(IsmType.WEIGHTED_MERKLE_ROOT_MULTISIG),
+      z.literal(IsmType.WEIGHTED_MESSAGE_ID_MULTISIG),
+    ]),
+  }),
+);
+
+export const RoutingIsmConfigSchema: z.ZodSchema<RoutingIsmConfig> = z.lazy(
+  () =>
+    z.discriminatedUnion('type', [
+      z.object({
+        type: z.literal(IsmType.ICA_ROUTING),
+      }),
+      OwnableSchema.extend({
+        type: z.literal(IsmType.ROUTING),
+        domains: z.record(IsmConfigSchema),
+      }),
+      OwnableSchema.extend({
+        type: z.literal(IsmType.FALLBACK_ROUTING),
+        domains: z.record(IsmConfigSchema),
+      }),
+    ]),
+);
+
+export const AggregationIsmConfigSchema: z.ZodSchema<AggregationIsmConfig> = z
+  .lazy(() =>
+    z.object({
+      type: z.literal(IsmType.AGGREGATION),
+      modules: z.array(IsmConfigSchema),
+      threshold: z.number(),
+    }),
+  )
+  .refine((data) => data.threshold <= data.modules.length, {
+    message: 'Threshold must be less than or equal to the number of modules',
+  });
+
+export const IsmConfigSchema = z.union([
+  ZHash,
+  TestIsmConfigSchema,
+  OpStackIsmConfigSchema,
+  PausableIsmConfigSchema,
+  TrustedRelayerIsmConfigSchema,
+  MultisigIsmConfigSchema,
+  WeightedMultisigIsmConfigSchema,
+  RoutingIsmConfigSchema,
+  AggregationIsmConfigSchema,
+  ArbL2ToL1IsmConfigSchema,
+]);