+++ typescript/utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hyperlane-xyz/utils
 
+## 7.3.0
+
+## 7.2.0
+
+### Minor Changes
+
+- fa6d5f5c6: Add toUpperCamelCase and deepFind functions
+
 ## 7.1.0
 
 ### Minor Changes